change the type of log.retension.size config from int to long

+++ src/main/java/com/leansoft/luxun/server/ServerConfig.java
@@ -93,8 +93,8 @@ public class ServerConfig {
     }
 
     /** the maximum total size of all log page files before deleting it */
-    public int getLogRetentionSize() {
-        return getInt(props, "log.retention.size", -1);
+    public long getLogRetentionSize() {
+        return getLong(props, "log.retention.size", -1L);
     }
 
     /**
